+++ ui-react/packages/atlasmap-provider/src/useAtlasmap.tsx
@@ -8,6 +8,45 @@ import { FileManagementService } from './services/file-management.service';
 import { InitializationService } from './services/initialization.service';
 import { MappingManagementService } from './services/mapping-management.service';
 
+
+export function importAtlasFile(selectedFile: File) {
+  console.log('importAtlasFile: ' + selectedFile.name);
+
+  const userFileComps = selectedFile.name.split('.');
+  const userFileSuffix: string = userFileComps[userFileComps.length - 1].toUpperCase();
+
+  if (userFileSuffix === 'ADM') {
+        /*
+    const error: any = null;
+    new ErrorInfo({
+      message: 'This is a test of the AtlasMap error service.',
+      level: ErrorLevel.ERROR, scope: ErrorScope.APPLICATION, type: ErrorType.INTERNAL, 
+      object: error});
+
+      this.cfg.errorService.resetAll();
+
+      // Clear out current user documents from the runtime service before processing the imported ADM.
+      this.cfg.fileService.resetMappings().toPromise().then( async() => {
+        this.cfg.fileService.resetLibs().toPromise().then( async() => {
+          await this.processMappingsCatalog(selectedFile);
+        });
+      }).catch((error: any) => {
+        if (error.status === 0) {
+          this.cfg.errorService.addError(new ErrorInfo({
+            message: 'Fatal network error: Could not connect to AtlasMap design runtime service.',
+            level: ErrorLevel.ERROR, scope: ErrorScope.APPLICATION, type: ErrorType.INTERNAL, object: error}));
+        } else {
+          this.cfg.errorService.addError(new ErrorInfo({
+            message: 'Could not reset document definitions before import.',
+            level: ErrorLevel.ERROR, scope: ErrorScope.APPLICATION, type: ErrorType.INTERNAL, object: error}));
+        }
+      });
+      */
+  } else if (userFileSuffix === 'JAR') {
+    // this.cfg.documentService.processDocument(selectedFile, InspectionType.JAVA_CLASS, false);
+  }
+}
+
 export interface IUseAtlasmapArgs {
   baseJavaInspectionServiceUrl: string;
   baseXMLInspectionServiceUrl: string;
@@ -64,5 +103,5 @@ export function useAtlasmap({
     baseMappingServiceUrl,
   ]);
 
-  return { sourceDocs, targetDocs };
+  return { sourceDocs, targetDocs, importAtlasFile };
 }
